feat: display user info on profile page

+++ utils/index.ts
@@ -1,5 +1,6 @@
 export * from './file'
 export * from './game'
+export * from './tags'
 export * from './user'
 
 export function isEmptyObj(obj: Record<string, unknown>): boolean {
@@ -9,14 +10,17 @@ export function isEmptyObj(obj: Record<string, unknown>): boolean {
 export function buildQuerySting(
   key: string,
   value?: string,
-  initial: Record<string, string> = {}
+  initial: Record<string, string | string[]> = {}
 ): string {
-  const query = new URLSearchParams(initial)
-  if (value) {
-    query.set(key, value)
-  } else {
-    query.delete(key)
+  const query = new URLSearchParams()
+  if (initial) {
+    for (const [key, value] of Object.entries(initial)) {
+      if (typeof value === 'string') query.append(key, value)
+      if (Array.isArray(value)) value.forEach((v) => query.append(key, v))
+    }
   }
+  if (value) query.set(key, value)
+  if (!value || value === '#') query.delete(key)
   const queryStr = query.toString()
   return queryStr ? `?${queryStr}` : ''
 }
